# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KPI Laravel API",
        "description": "Multi-tenant HR KPI platform. Covers v1 (existing) and v2 (multi-tenant) endpoints.",
        "contact": {
            "email": "admin@system.local"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local / Development"
        }
    ],
    "paths": {
...
        "/api/v2/tenants": {
            "get": {
                "tags": [
                    "Tenants"
                ],
                "summary": "List all tenants (paginated)",
                "operationId": "671572354240e95ced976177b8636c54",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "suspended"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated tenant list"
                    },
                    "403": {
                        "description": "Forbidden — super_admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Create a new tenant",
                "operationId": "f546e35f79b5090a5c125a950235dbb7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tenant_code",
                                    "tenant_name"
                                ],
                                "properties": {
                                    "tenant_code": {
                                        "type": "string",
                                        "example": "COMP-001"
                                    },
                                    "tenant_name": {
                                        "type": "string",
                                        "example": "PT Example"
                                    },
                                    "domain": {
                                        "type": "string",
                                        "example": "app.example.com"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "suspended"
                                        ]
                                    },
                                    "contact_email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "contact_phone": {
                                        "type": "string",
                                        "example": "+628123456789"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tenant created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/tenants/{id}": {
            "get": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Get a single tenant with counts",
                "operationId": "44a0a98f22695be71879b8dbfd782057",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant detail"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Update a tenant",
                "operationId": "d02cbca7264081b1cf0524e4f904f0aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tenant_name": {
                                        "type": "string"
                                    },
                                    "domain": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "suspended"
                                        ]
                                    },
                                    "contact_email": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tenant updated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/tenants/{id}/activate": {
            "patch": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Activate a tenant",
                "operationId": "5837fb8ca44c9dfb43542e2d4f48373f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant activated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/tenants/{id}/deactivate": {
            "patch": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Deactivate a tenant",
                "operationId": "d4dfc262a69713acce693f2df018e636",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant deactivated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Include: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}